Add consolidation of field validations
