+++ src/js/common/utils/validations.js
@@ -1,22 +1,14 @@
 import _ from 'lodash';
 import moment from 'moment';
 import { states } from './options-for-select';
+import { dateToMoment } from './helpers';
 
+/**
+ * General Validations *
+*/
 function validateIfDirty(field, validator) {
   if (field.dirty) {
     return validator(field.value);
-  }
-
-  return true;
-}
-
-function isDirtyDate(date) {
-  return date.day.dirty && date.year.dirty && date.month.dirty;
-}
-
-function validateIfDirtyDate(dayField, monthField, yearField, validator) {
-  if (isDirtyDate({ day: dayField, month: monthField, year: yearField })) {
-    return validator(dayField.value, monthField.value, yearField.value);
   }
 
   return true;
@@ -44,43 +36,39 @@ function isNotBlank(value) {
   return value !== '';
 }
 
-function isNotBlankDateField(field) {
-  return isNotBlank(field.day.value) && isNotBlank(field.month.value) && isNotBlank(field.year.value);
+function isValidValue(validator, value) {
+  return isBlank(value) || validator(value);
 }
 
-// Conditions for valid SSN from the original 1010ez pdf form:
-// '123456789' is not a valid SSN
-// A value where the first 3 digits are 0 is not a valid SSN
-// A value where the 4th and 5th digits are 0 is not a valid SSN
-// A value where the last 4 digits are 0 is not a valid SSN
-// A value with 3 digits, an optional -, 2 digits, an optional -, and 4 digits is a valid SSN
-// 9 of the same digits (e.g., '111111111') is not a valid SSN
-function isValidSSN(value) {
-  if (value === '123456789' || value === '123-45-6789') {
-    return false;
-  } else if (/^0{3}-?\d{2}-?\d{4}$/.test(value)) {
-    return false;
-  } else if (/^\d{3}-?0{2}-?\d{4}$/.test(value)) {
-    return false;
-  } else if (/^\d{3}-?\d{2}-?0{4}$/.test(value)) {
-    return false;
-  }
-
-  const noBadSameDigitNumber = _.without(_.range(0, 10), 2, 4, 5)
-    .every(i => {
-      const sameDigitRegex = new RegExp(`${i}{3}-?${i}{2}-?${i}{4}`);
-      return !sameDigitRegex.test(value);
-    });
-
-  if (!noBadSameDigitNumber) {
-    return false;
-  }
-
-  return /^\d{3}-?\d{2}-?\d{4}$/.test(value);
+function isValidField(validator, field) {
+  return isBlank(field.value) || validator(field.value);
 }
 
+function isValidRequiredField(validator, field) {
+  return isNotBlank(field.value) && validator(field.value);
+}
+
+/**
+ * Date Validations *
+*/
 function isValidYear(value) {
   return Number(value) >= 1900 && Number(value) <= moment().add(100, 'year').year();
+}
+
+function isValidYearOrBlank(value) {
+  return isValidYear(value) || value === '';
+}
+
+function isValidCurrentOrPastYear(value) {
+  return Number(value) >= 1900 && Number(value) < moment().year() + 1;
+}
+
+function isValidMonths(value) {
+  return Number(value) >= 0;
+}
+
+function isBlankDateField(field) {
+  return isBlank(field.day.value) && isBlank(field.month.value) && isBlank(field.year.value);
 }
 
 function isValidDate(day, month, year) {
@@ -103,6 +91,22 @@ function isValidDate(day, month, year) {
     date.getFullYear() === Number(year);
 }
 
+function isNotBlankDateField(field) {
+  return isNotBlank(field.day.value) && isNotBlank(field.month.value) && isNotBlank(field.year.value);
+}
+
+function isDirtyDate(date) {
+  return date.day.dirty && date.year.dirty && date.month.dirty;
+}
+
+function validateIfDirtyDate(dayField, monthField, yearField, validator) {
+  if (isDirtyDate({ day: dayField, month: monthField, year: yearField })) {
+    return validator(dayField.value, monthField.value, yearField.value);
+  }
+
+  return true;
+}
+
 function isValidAnyDate(day, month, year) {
   if (!isValidYear(year)) {
     return false;
@@ -123,6 +127,24 @@ function isValidPartialDate(day, month, year) {
   return true;
 }
 
+function isValidDateField(field) {
+  return isValidDate(field.day.value, field.month.value, field.year.value);
+}
+
+function isValidPartialDateField(field) {
+  return isValidPartialDate(field.day.value, field.month.value, field.year.value);
+}
+
+function isValidDateRange(fromDate, toDate) {
+  if (isBlankDateField(toDate) || isBlankDateField(fromDate)) {
+    return true;
+  }
+  const momentStart = dateToMoment(fromDate);
+  const momentEnd = dateToMoment(toDate);
+
+  return momentStart.isBefore(momentEnd);
+}
+
 function isValidPartialMonthYear(month, year) {
   if (typeof month === 'object') {
     throw new Error('Pass a month and a year to function');
@@ -132,6 +154,16 @@ function isValidPartialMonthYear(month, year) {
   }
 
   return isValidPartialDate(null, null, year);
+}
+
+function isValidPartialMonthYearRange(fromDate, toDate) {
+  if (!fromDate.year.value || !toDate.year.value) {
+    return true;
+  }
+  const momentStart = dateToMoment(fromDate);
+  const momentEnd = dateToMoment(toDate);
+
+  return momentStart.isSameOrBefore(momentEnd);
 }
 
 function isValidPartialMonthYearInPast(month, year) {
@@ -154,88 +186,6 @@ function isValidDateOver17(day, month, year) {
     year
   });
   return momentDate.isBefore(moment().endOf('day').subtract(17, 'years'));
-}
-
-function isValidName(value) {
-  return /^[a-zA-Z][a-zA-Z '\-]*$/.test(value);
-}
-
-function isValidMonetaryValue(value) {
-  if (value !== null) {
-    return /^\d+\.?\d*$/.test(value);
-  }
-  return true;
-}
-
-function isValidUSZipCode(value) {
-  return /(^\d{5}$)|(^\d{5}[ -]{0,1}\d{4}$)/.test(value);
-}
-
-function isValidCanPostalCode(value) {
-  return /^[a-zA-Z]\d[a-zA-Z][ -]{0,1}\d[a-zA-Z]\d$/.test(value);
-}
-
-// TODO: look into validation libraries (npm "validator")
-function isValidPhone(value) {
-  // Strip spaces, dashes, and parens
-  const stripped = value.replace(/[^\d]/g, '');
-  // Count number of digits
-  return /^\d{10}$/.test(stripped);
-}
-
-function isValidEmail(value) {
-  // Comes from StackOverflow: http://stackoverflow.com/questions/46155/validate-email-address-in-javascript
-  return /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(value);
-}
-
-function isValidField(validator, field) {
-  return isBlank(field.value) || validator(field.value);
-}
-
-function isValidRequiredField(validator, field) {
-  return isNotBlank(field.value) && validator(field.value);
-}
-
-function isBlankDateField(field) {
-  return isBlank(field.day.value) && isBlank(field.month.value) && isBlank(field.year.value);
-}
-
-function isValidDateField(field) {
-  return isValidDate(field.day.value, field.month.value, field.year.value);
-}
-
-function isValidPartialDateField(field) {
-  return isValidPartialDate(field.day.value, field.month.value, field.year.value);
-}
-
-function isValidFullNameField(field) {
-  return isValidName(field.first.value) &&
-    (isBlank(field.middle.value) || isValidName(field.middle.value)) &&
-    isValidName(field.last.value);
-}
-
-function isValidAddressField(field) {
-  const initialOk = isNotBlank(field.street.value) &&
-    isNotBlank(field.city.value) &&
-    isNotBlank(field.country.value);
-
-  // if we have a defined list of values, they will
-  // be set as the state and zipcode keys
-  if (_.hasIn(states, field.country.value)) {
-    return initialOk &&
-      isNotBlank(field.state.value) &&
-      isNotBlank(field.zipcode.value);
-  }
-  // if the entry was non-USA/CAN/MEX, only postal is
-  // required, not provinceCode
-  return initialOk && isNotBlank(field.postalCode.value);
-}
-
-function isBlankAddress(address) {
-  return isBlank(address.city.value)
-    && isBlank(address.state.value)
-    && isBlank(address.street.value)
-    && isBlank(address.postalCode.value);
 }
 
 function isValidEntryDateField(date, dateOfBirth) {
@@ -274,37 +224,138 @@ function isValidDischargeDateField(date, entryDate) {
   return true;
 }
 
+/**
+ * Field Validations *
+*/
+function isValidName(value) {
+  return /^[a-zA-Z][a-zA-Z '\-]*$/.test(value);
+}
+
+function isValidFullNameField(field) {
+  return isValidName(field.first.value) &&
+    (isBlank(field.middle.value) || isValidName(field.middle.value)) &&
+    isValidName(field.last.value);
+}
+
+// Conditions for valid SSN from the original 1010ez pdf form:
+// '123456789' is not a valid SSN
+// A value where the first 3 digits are 0 is not a valid SSN
+// A value where the 4th and 5th digits are 0 is not a valid SSN
+// A value where the last 4 digits are 0 is not a valid SSN
+// A value with 3 digits, an optional -, 2 digits, an optional -, and 4 digits is a valid SSN
+// 9 of the same digits (e.g., '111111111') is not a valid SSN
+function isValidSSN(value) {
+  if (value === '123456789' || value === '123-45-6789') {
+    return false;
+  } else if (/^0{3}-?\d{2}-?\d{4}$/.test(value)) {
+    return false;
+  } else if (/^\d{3}-?0{2}-?\d{4}$/.test(value)) {
+    return false;
+  } else if (/^\d{3}-?\d{2}-?0{4}$/.test(value)) {
+    return false;
+  }
+
+  const noBadSameDigitNumber = _.without(_.range(0, 10), 2, 4, 5)
+    .every(i => {
+      const sameDigitRegex = new RegExp(`${i}{3}-?${i}{2}-?${i}{4}`);
+      return !sameDigitRegex.test(value);
+    });
+
+  if (!noBadSameDigitNumber) {
+    return false;
+  }
+
+  return /^\d{3}-?\d{2}-?\d{4}$/.test(value);
+}
+
+function isValidMonetaryValue(value) {
+  if (value !== null) {
+    return /^\d+\.?\d*$/.test(value);
+  }
+  return true;
+}
+
+function isValidUSZipCode(value) {
+  return /(^\d{5}$)|(^\d{5}[ -]{0,1}\d{4}$)/.test(value);
+}
+
+function isValidCanPostalCode(value) {
+  return /^[a-zA-Z]\d[a-zA-Z][ -]{0,1}\d[a-zA-Z]\d$/.test(value);
+}
+
+// TODO: look into validation libraries (npm "validator")
+function isValidPhone(value) {
+  // Strip spaces, dashes, and parens
+  const stripped = value.replace(/[^\d]/g, '');
+  // Count number of digits
+  return /^\d{10}$/.test(stripped);
+}
+
+function isValidEmail(value) {
+  // Comes from StackOverflow: http://stackoverflow.com/questions/46155/validate-email-address-in-javascript
+  return /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(value);
+}
+
+function isValidAddressField(field) {
+  const initialOk = isNotBlank(field.street.value) &&
+    isNotBlank(field.city.value) &&
+    isNotBlank(field.country.value);
+
+  // if we have a defined list of values, they will
+  // be set as the state and zipcode keys
+  if (_.hasIn(states, field.country.value)) {
+    return initialOk &&
+      isNotBlank(field.state.value) &&
+      isNotBlank(field.zipcode.value);
+  }
+  // if the entry was non-USA/CAN/MEX, only postal is
+  // required, not provinceCode
+  return initialOk && isNotBlank(field.postalCode.value);
+}
+
+function isBlankAddress(address) {
+  return isBlank(address.city.value)
+    && isBlank(address.state.value)
+    && isBlank(address.street.value)
+    && isBlank(address.postalCode.value);
+}
+
 export {
-  validateIfDirty,
-  validateIfDirtyDate,
   isBlank,
   isBlankDateField,
+  isBlankAddress,
+  isDirtyDate,
   isNotBlank,
+  isNotBlankDateField,
   isValidAddressField,
-  isValidRequiredField,
+  isValidAnyDate,
+  isValidCanPostalCode,
+  isValidCurrentOrPastYear,
   isValidDate,
   isValidDateField,
-  isValidName,
-  isValidSSN,
-  isValidMonetaryValue,
-  isValidUSZipCode,
-  isValidCanPostalCode,
-  isValidPhone,
+  isValidDateOver17,
+  isValidDateRange,
+  isValidDischargeDateField,
   isValidEmail,
   isValidEntryDateField,
-  isValidDischargeDateField,
   isValidFullNameField,
   isValidField,
-  isBlankAddress,
-  isValidAnyDate,
-  isValidDateOver17,
-  isDirtyDate,
-  isNotBlankDateField,
+  isValidMonths,
+  isValidName,
+  isValidMonetaryValue,
+  isValidPhone,
   isValidPartialDate,
-  isValidDateField,
   isValidPartialDateField,
   isValidPartialMonthYear,
-  isValidYear,
   isValidPartialMonthYearInPast,
-  validateCustomFormComponent
+  isValidPartialMonthYearRange,
+  isValidRequiredField,
+  isValidSSN,
+  isValidUSZipCode,
+  isValidValue,
+  isValidYear,
+  isValidYearOrBlank,
+  validateCustomFormComponent,
+  validateIfDirty,
+  validateIfDirtyDate,
 };
